requirements.txt
fastapi==0.116.1
httpx==0.28.1
psycopg[binary]==3.2.9
python-dotenv==1.0.1
uvicorn==0.35.0
